+++ src/components/installation-section/GettingStarted.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import InstallationItem1 from './InstallationItems/InstallationItem1.vue';
-import InstallationItem3 from './InstallationItems/InstallationItem2.vue';
+import Installation from './startingItems/Installation.vue';
+import Examples from './startingItems/Examples.vue';
+import Documentation from './startingItems/Documentation.vue';
 </script>
 
 <template>
@@ -10,11 +11,12 @@ import InstallationItem3 from './InstallationItems/InstallationItem2.vue';
       <h1
         class="py-8 text-4xl bg-gradient-to-r from-greenP to-whiteB text-transparent bg-clip-text"
       >
-        Installation
+        Getting Started
       </h1>
       <div class="flex-col">
-        <InstallationItem1 />
-        <InstallationItem3 />
+        <Installation />
+        <Examples />
+        <Documentation />
       </div>
     </div>
   </section>
