+++ src/components/installation-section/startingItems/Examples.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts"></script>
 
 <template>
-  <h1 class="py-4 text-2xl text-whiteB">Code Sandbox</h1>
+  <h1 class="py-4 text-2xl text-whiteB">Examples</h1>
   <p class="text-whiteB font-light">
     placeholder text to see how this will look when it's more filled out
     placeholder text to see how this will look when it's more filled out
