+++ src/components/installation-section/startingItems/Installation.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts"></script>
 
 <template>
-  <h1 class="py-4 text-2xl text-whiteB">Getting Started</h1>
+  <h1 class="py-4 text-2xl text-whiteB">Installation</h1>
   <p class="text-whiteB font-light">
     placeholder text to see how this will look when it's more filled out
     placeholder text to see how this will look when it's more filled out
